requirements.txt
numpy
scipy
pillow
imageio
PyYAML
matplotlib
